requirements.txt
langchain
langchain-huggingface
langsmith
python-dotenv
transformers
torch
accelerate